requirements.txt
docx2txt
fuzzywuzzy
uvicorn
pdfminer.six